Added redirects and functions to prevent users from manually entering urls to bypass credential checks, added rank to session to track admin privileges.

+++ run.py
@@ -32,9 +32,25 @@ def send_file(filename):
 
 # Helper functions for routes
 
+# Checks if the user owns the topic/post or is admin, shows message to user
+# if 'False'
+def check_owner_or_admin(author):
+    if author == session['user_id'] or session['rank'] == 'admin':
+        return True
+    else:
+        flash("You are not the owner.")
+        return False
+
+
 # Grabs topics collection from db
 def fetch_topics():
     return list(mongo.db.topics.find())
+
+
+# Grabs single topic based on the id
+def fetch_single_topic(topic_id):
+    topic_info = mongo.db.topics.find_one({'_id': ObjectId(topic_id)})
+    return topic_info
 
 
 # Checks if user is logged in
@@ -80,15 +96,18 @@ def update_topic_document(topic_info):
 # Creates update topic document, update db and flashes message confirming the
 # update
 def update_topic(topic_info):
-    submit = update_topic_document(topic_info)
-    mongo.db.topics.update({"_id": topic_info["_id"]}, submit)
-    flash("Topic Successfully Updated")
+    if check_owner_or_admin(topic_info['_id']):
+        submit = update_topic_document(topic_info)
+        mongo.db.topics.update({"_id": topic_info["_id"]}, submit)
+        flash("Topic Successfully Updated")
 
 
 # Deletes topic from db and flashes message confirming deletion
-def remove_topic(topic):
-    mongo.db.topics.remove({"_id": ObjectId(topic)})
-    flash("Topic has been deleted.")
+def remove_topic(topic_id):
+    topic_info = fetch_single_topic(topic_id)
+    if check_owner_or_admin(topic_info['_id']):
+        mongo.db.topics.remove({"_id": ObjectId(topic_id)})
+        flash("Topic has been deleted.")
 
 
 # Fetch topic from db and all associated posts
@@ -152,18 +171,20 @@ def update_post_document(post_info):
 
 # Updates post and flashes message confirming changes
 def update_post(post_info):
-    post = update_post_document(post_info)
-    mongo.db.posts.update({"_id": post_info['_id']}, post)
-    flash("Post Successfully Updated")
+    if check_owner_or_admin(post_info['_id']):
+        post = update_post_document(post_info)
+        mongo.db.posts.update({"_id": post_info['_id']}, post)
+        flash("Post Successfully Updated")
 
 
 # Modifies post count for topic and user, deletes post and flashes message
 # confirming changes
 def remove_post(post):
-    modify_post_count(post['topic'], post['author'], -1)
-    # Delete post from db and flash message confirming deletion
-    mongo.db.posts.delete_one({"_id": ObjectId(post['_id'])})
-    flash("Post has been deleted.")
+    if check_owner_or_admin(post):
+        modify_post_count(post['topic'], post['author'], -1)
+        # Delete post from db and flash message confirming deletion
+        mongo.db.posts.delete_one({"_id": ObjectId(post['_id'])})
+        flash("Post has been deleted.")
 
 
 # Saves image in upload folder and grabs filename to save in db
@@ -189,11 +210,12 @@ def update_user_document(user):
 # Saves profile image, updates user in db, updates cookies and shows
 # message confirming update
 def update_user(user):
-    save_profile_image()
-    submit = update_user_document(user)
-    mongo.db.users.update({"_id": ObjectId(user['_id'])}, submit)
-    session['display_name'] = submit['display_name']
-    flash("Profile Successfully Updated")
+    if check_owner_or_admin(user['_id']):
+        save_profile_image()
+        submit = update_user_document(user)
+        mongo.db.users.update({"_id": ObjectId(user['_id'])}, submit)
+        session['display_name'] = submit['display_name']
+        flash("Profile Successfully Updated")
 
 
 # Fetch user from db
@@ -227,8 +249,8 @@ def set_profile_picture(user_id):
 # message confirming registration
 def set_cookies(register, user_id):
     session['display_name'] = register['display_name']
+    session['rank'] = register['rank']
     session['user_id'] = user_id
-    flash("Registration Successful!")
 
 
 # Inserts new user into db, sets default profile picture and sets cookies
@@ -239,6 +261,7 @@ def register_user():
         {"email": register["email"]})["_id"])
     set_profile_picture(user_id)
     set_cookies(register, user_id)
+    flash("Registration Successful!")
 
 
 # Logs user out, shows message confirming changes and
@@ -247,6 +270,7 @@ def clear_session():
     flash("You have been logged out.")
     session.pop("display_name")
     session.pop("user_id")
+    session.pop("rank")
 
 
 # Queries the topics collection according to the value entered
@@ -264,6 +288,7 @@ def index():
 
     # Fetch topics collection from db
     topics = fetch_topics()
+    print(session['rank'])
 
     return render_template("index.html", topics=topics)
 
@@ -285,8 +310,11 @@ def index_post():
 @app.route("/edit_topic/<topic>", methods=["POST"])
 def edit_topic(topic):
 
+    if not check_user_is_logged_in():
+        return redirect(url_for("login"))
+
     # Fetch topic info from db
-    topic_info = mongo.db.topics.find_one({'_id': ObjectId(topic)})
+    topic_info = fetch_single_topic(topic)
 
     # Update topic and flash message confirming changes
     update_topic(topic_info)
@@ -296,6 +324,9 @@ def edit_topic(topic):
 
 @app.route("/delete_topic/<topic>")
 def delete_topic(topic):
+
+    if not check_user_is_logged_in():
+        return redirect(url_for("login"))
 
     # Remove topic from db and flash message confirming deletion
     remove_topic(topic)
@@ -382,9 +413,8 @@ def edit_profile_post(user_id):
     # Fetch user from db
     user = fetch_user(user_id)
 
-    # Store profile image in db
-    # Update user and flash message confirming changes, then redirects
-    # to updated profile page
+    # Stores profile image in db, updates user and flashes message
+    # confirming changes, then redirects to updated profile page
     update_user(user)
 
     return redirect(url_for("profile", user_id=user_id))
@@ -400,6 +430,10 @@ def register_post():
 
     # Check if username already exists in db
     existing_user = check_user_registered()
+
+    # If session exists, redirects user to profile
+    if check_user_is_logged_in():
+        return redirect(url_for("profile", user_id=session['user_id']))
 
     # Shows message to user confirming that account already
     # exists and redirects to register
@@ -422,6 +456,10 @@ def login_post():
 
     # Check if usernarme exists in db
     existing_user = check_user_registered()
+
+    # If session exists, redirects user to profile
+    if check_user_is_logged_in():
+        return redirect(url_for("profile", user_id=session['user_id']))
 
     if existing_user:
         # Ensures hashed password matches user input, sets cookies,
